refactor: extract the backtest candle source
All Saxo fetching and DynamoDB raw-candle caching moves into
CandleSource, the only module in the package that performs I/O. It
answers one question - what candles does this definition see on this
day - and returns None when the day has no usable data.

evaluate_day drops from ~110 lines of interleaved cache lookup, cache
shape validation, two try/except fetch paths and three store calls to
six lines: get the candles, evaluate them. The wide-range filter is no
longer evaluated twice on the miss path; the predicate itself moves to
definitions.py, where both the fetch short-circuit and the strategy read
it from one place.
